requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
PyGithub==2.4.0
anthropic==0.34.0
langchain==0.3.0
langgraph==0.2.28
langchain-anthropic==0.2.0
langsmith==0.1.117
pydantic==2.9.0
pydantic-settings==2.5.2
httpx==0.27.2
celery==5.4.0
redis==5.1.0
pytest==8.3.0
pytest-asyncio==0.24.0